Major: move to a 2^128 file id space.
The long random id used to identify each file is replaced by an instance of
FileId. Currently, a FileId consists in two longs, which increases the state
space to 2^128. This should be more than enough
(see https://en.wikipedia.org/wiki/Birthday_attack), but this is now modifiable
if needed as the FileId is mostly opaque.