Allow the usage of generic and nested transfer plugins
This version allows to use generic plugins as nested ones. The user is
asked which plugin he wants to use for e.g. the creation of a raid. It
also supports the non-interactive initialization of plugins using the
new, magic parameter <fieldname>-tpt as plugin parameter.

TransferPluginUtil now offers a way to find the transfer plugin
implementation for a transfer settings class.

Furthermore this commit fixes a bug when setting null values as
setting values.