Added postQueryCallback and PluginOptionConverter
postQueryCallback is counterpart of preQueryCallback and is trigged
after a user entered some a setting value.

PluginOptionConverters are used to convert a user input value after a
he entered it. Converters become necessary if transfer plugins use an
oauth token, for instance. Such plugins need to validate and extract
information from a received token.