Bugfixes and singular field values
Fixed a bug that encrypted fields in transfer settings were encrypted or
decrypted multiple times. This resulted in the link which was generated
while init being encrypted twice.

Added singular field parameter to setup annotation. Singular fields have
to be entered every time a value is queried. For instance, if a user
revises setup values during init singular fields are not cached. This is
needed for plugins which do oauth since access tokens might be
invalidated after the first call (this is case with the dropbox plugin).