OSX launchdaemon and homebrew formula
The OSX launchdaemon has to be moved to /Library/LaunchDaemons/. As a
result the syncany daemon will be launched once the system is started.
TODO: move the file using gradle to the specified directory.

The homebrew formula allows less experienced users to use syncany
without having to manually build it. Homebrew only accepts production
ready software, therefore we need to provide the formula. (thanks to
@tvon, closes syncany/syncany#267).