Wrap TransferManagers to extend features, Attributes for RemoteFile
A basic TM is extendable by annotating a class with features.
TransferManagerFeatureSupport provides a builder to wrap such
TransferManagers with a defined set of supported features.

Options for specific TMs can be set using the same annotation which are
used to enable a feature.

RemoteFiles can now hold a specific, TM defined, set of attributes. For
instance, PathAware adds a path attribute to RemoteFiles without
polluting the basic implementation.