Cleaning up code.
Moved some code around. In particular the checks for when to stop
indexing/transactions.

Also provided a few 'Listener' interfaces which make the structure we use for
filling a queue with DatabaseVersions much cleaner. These interfaces may need to
be placed in a util package. Would like to discuss this.