Add OauthTokenWebListener
OauthTokenWebListener makes it unnecessary to copy and paste an token
from a url while performing the oauth process. It opens an server which
can be used as a callback URL during oauth.
It currently supports extracting the token from a callback URL as a
query parameter but other extractors are wip.